requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ftfy
regex
git+https://github.com/openai/CLIP.git
tqdm
numpy
opencv-python
Pillow
pyyaml